Added hw07 files
Added initial hw files with task. Сonnected script and styles of the basicLightbox modal window library in 01-gallery.html and сompleted the task of creating and rendering the galleryItems data array markup and the provided gallery element template in 01-gallery.js. In task 2, connected the script and styles of the SimpleLightbox library in 02-lightbox.html. Also, used the ready-made code from the first task to create and render the galleryItems data array markup and the gallery element template in 02-lightbox.js